+++ action_server/src/robocorp/action_server/_settings.py
@@ -8,7 +8,7 @@ from typing import Iterator, Optional
 
 from termcolor import colored
 
-from ._protocols import ArgumentsNamespace
+from ._protocols import ArgumentsNamespaceMigrateImportOrStart
 
 log = logging.getLogger(__name__)
 
@@ -85,7 +85,7 @@ class Settings:
         return ret
 
     @classmethod
-    def create(cls, args: ArgumentsNamespace) -> "Settings":
+    def _create(cls, args: ArgumentsNamespaceMigrateImportOrStart) -> "Settings":
         user_specified_datadir = args.datadir
         if not user_specified_datadir:
             import hashlib
@@ -146,9 +146,11 @@ _global_settings: Optional[Settings] = None
 
 
 @contextmanager
-def setup_settings(args: ArgumentsNamespace) -> Iterator[Settings]:
+def setup_settings(
+    args: ArgumentsNamespaceMigrateImportOrStart,
+) -> Iterator[Settings]:
     global _global_settings
-    settings = Settings.create(args)
+    settings = Settings._create(args)
     _global_settings = settings
     try:
         yield settings