+++ log_pytest/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Update package's main README.
+
 ## 0.0.3 - 2024-01-14
 
 - Fix main README and update docs.