+++ workitems/docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Update package's main README.
+
 ## 1.4.4 - 2024-03-19
 
 - Enables compatibility with Action Server `0.1.0` by accepting **robocorp-tasks**